#608 refactor search results empty screen
