+++ src/screens/search_results/index.ts
@@ -1,0 +1,4 @@
+import SearchResultsScreen from './SearchResultsScreen';
+import EmptySearchResults from './EmptySearchResults';
+
+export { SearchResultsScreen, EmptySearchResults };
